Taller 2

+++ Entidades/EstadoCivil.cs
@@ -1,7 +1,7 @@
-﻿namespace Entidades
+﻿ namespace Entidades
 {
     public class EstadoCivil
-    { public int Id { get; set;}
+    { public short Id { get; set;}
         public string Nombre { get; set; }
     }
 }
